% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark_db}
\alias{benchmark_db}
\title{Benchmark database(s) against a given function}
\usage{
benchmark_db(connections, f, ...)
}
\arguments{
\item{connections}{A named list of DBI connections}

\item{f}{Function to benchmark. Should take DBI connection as an argument}

\item{...}{Arguments passed to microbenchmark}
}
\description{
Benchmark database(s) against a given function
}
